作成完了

+++ app/src/main/java/jp/techacademy/yuuta/calcapp/SecondActivity.kt
@@ -10,7 +10,7 @@ class SecondActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_second)
 
-        val value1 = intent.getIntExtra("VALUE1", 0)
+        val value1 = intent.getDoubleExtra("VALUE1", 0.0)
 
         textView.text = "${value1}"
     }
